backend/requirements.txt
django
requests
numpy
pandas
lark
scipy
sympy
djangorestframework
django-cors-headers
qiskit
qiskit-optimization
qiskit-ibmq-provider
qiskit-ibm-runtime

matplotlib
pylatexenc